+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- Nothing yet!
+### Fixed
+
+- Fix infinite loop when `theme()` function is called recursively inside `@theme` blocks
 
 ## [4.1.6] - 2025-05-09